// src/infrastructure/persistence/supabase/types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      clients: {
        Row: {
          created_at: string
          display_number: number
          entity_id: string
          id: string
          instagram: string | null
          lead_source: string | null
          name: string
          phone: string | null
          updated_at: string
        }
        Insert: {
          created_at?: string
          display_number: number
          entity_id: string
          id?: string
          instagram?: string | null
          lead_source?: string | null
          name: string
          phone?: string | null
          updated_at?: string
        }
        Update: {
          created_at?: string
          display_number?: number
          entity_id?: string
          id?: string
          instagram?: string | null
          lead_source?: string | null
          name?: string
          phone?: string | null
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "clients_entity_id_fkey"
            columns: ["entity_id"]
            isOneToOne: false
            referencedRelation: "entity"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}